♻️ refactor: pathfindingモジュールの構造化とファイル分割
- pathfinding.tsを機能別に複数ファイルに分割
- src/logic/pathfinding/フォルダを作成して関連ファイルを整理
- BFS探索アルゴリズム（bfs.ts）とグリッド変換処理を分離
- クリア時足跡描画（footprint-grid.ts）とRest状態変更（rest-transition.ts）を独立
- メイン経路探索関数（main.ts）でコア機能を集約
- 適切なindex.tsでエクスポートを統一
