+++ src/logic/pathfinding/index.ts
@@ -1,0 +1,9 @@
+// パズル経路探索のメイン関数
+export { findPath } from './main';
+
+// BFS探索アルゴリズム
+export { bfsAllShortestPaths } from './bfs';
+
+// グリッド状態変更関数
+export { createFootprintGrid } from './footprint-grid';
+export { createRestTransitionGrid } from './rest-transition';
